bot/requirements.txt
asyncpg
pillow
numpy